+++ Frontend/src/Components/ScrollUp.css
@@ -1,4 +1,3 @@
-/* Existing scroll button styles */
 .scroll-up-btn {
   position: fixed;
   display: flex;
@@ -6,12 +5,13 @@
   align-items: center;
   bottom: 35px;
   right: 25px;
-  width: 45px;
-  height: 45px;
+  width: 45px ;
+  height: 45px ;
   z-index: 1000;
-  background-color: white;
-  color: #ff8e3a;
-  border: none;
+    background-color: white ; 
+  color: #ff8e3a ;
+    border: none;
+
   border-radius: 50%;
   cursor: pointer;
   font-size: 20px;
@@ -21,288 +21,4 @@
 
 .scroll-up-btn:hover {
   background-color: #555;
-}
-
-/* Button styles */
-.appluBtn {
-  position: fixed;
-  margin: 0;
-  padding: 6px 13px 8px 13px;
-  text-align: center;
-  font-size: 15px;
-  font-weight: bold;
-  border-style: solid;
-  display: block;
-  z-index: 100000;
-  background-color: #821d34;
-  margin-top: 200px;
-}
-
-.appluBtn_right {
-  right: -1px;
-  cursor: pointer;
-  transform: rotate(-90deg);
-  border-top-right-radius: 10px;
-  border-top-left-radius: 10px;
-}
-
-.appluBtn:hover {
-  background-color: #821d34;
-}
-
-.enquire-now-btns {
-  display: block;
-  padding: 5px 32px;
-  background-color: #821d34;
-  font-size: 15px;
-  text-align: center;
-  border-radius: 5px 5px 0px 0px;
-  border: none;
-  cursor: pointer;
-  font: inherit;
-  font-family: inherit;
-  font-style: inherit;
-  font-weight: inherit;
-  font-weight: bold;
-}
-
-.vertcalview {
-  position: fixed;
-  right: -70px;
-  top: 420px;
-  z-index: 999 !important;
-  transform: rotate(270deg);
-  color: #fff !important;
-}
-
-.vertcalview-1 {
-  position: fixed;
-  right: -38px;
-  top: 35px;
-  z-index: 999 !important;
-  transform: rotate(270deg);
-  color: #fff !important;
-}
-
-/* Popup styles */
-.popup-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0,0,0,0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 2000;
-}
-
-.popup-container {
-  position: relative;
-  max-width: 90%;
-  max-height: 90vh;
-  overflow-y: auto;
-  margin: 20px;
-}
-
-.popup-form {
-  background: #fff;
-  width: 100%;
-  max-width: 500px;
-  padding: 30px;
-  box-shadow: 0 0 15px rgba(0, 0, 0, 0.1);
-  border-radius: 8px;
-  position: relative;
-  font-family: 'Segoe UI', sans-serif;
-}
-
-.popup-form h3 {
-  font-size: 22px;
-  font-weight: 600;
-  margin-bottom: 20px;
-  text-align: center;
-}
-
-/* Form layout styles */
-.popup-form-content {
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
-
-/* Full width fields (Name, Email, Phone) */
-.popup-form input[type="text"],
-.popup-form input[type="email"],
-.popup-form input[type="tel"],
-.popup-form input[type="password"] {
-  width: 100%;
-  padding: 12px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  font-size: 14px;
-}
-
-/* Two-column layout for State, City, Program, Course */
-.form-row-1 {
-  display: flex;
-  gap: 15px;
-}
-
-.form-field {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-}
-
-.popup-form select {
-  width: 100%;
-  padding: 12px;
-  margin-bottom: 0;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  font-size: 14px;
-  background-color: white;
-  appearance: none;
-  -webkit-appearance: none;
-  background-repeat: no-repeat;
-  background-position: right 10px center;
-  background-size: 1em;
-}
-
-.popup-form input:focus,
-.popup-form select:focus {
-  border-color: #ff8000;
-  outline: none;
-  box-shadow: 0 0 0 2px rgba(255, 142, 58, 0.2);
-}
-
-/* Improved Checkbox Styles */
-.checkbox-row {
-  display: flex;
-  align-items: flex-start;
-  gap: 10px;
-  margin: 15px 0;
-}
-
-.checkbox-row input[type="checkbox"] {
-  appearance: none;
-  -webkit-appearance: none;
-  width: 18px;
-  height: 18px;
-  border: 2px solid #ff8e3a;
-  border-radius: 4px;
-  outline: none;
-  cursor: pointer;
-  position: relative;
-  flex-shrink: 0;
-  margin-top: 3px;
-}
-
-.checkbox-row input[type="checkbox"]:checked {
-  background-color: #ff8e3a;
-}
-
-.checkbox-row input[type="checkbox"]:checked::after {
-  content: "✓";
-  position: absolute;
-  color: white;
-  font-size: 12px;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.checkbox-row label {
-  font-size: 14px;
-  color: #555;
-  cursor: pointer;
-  line-height: 1.4;
-}
-
-.checkbox-row label:hover {
-  color: #333;
-}
-
-.popup-form .close-btn {
-  position: absolute;
-  top: 10px;
-  right: 15px;
-  font-size: 20px;
-  color: #333;
-  cursor: pointer;
-  background: none;
-  border: none;
-  padding: 5px;
-}
-
-.popup-form .close-btn:hover {
-  color: #ff8e3a;
-}
-
-/* Form Error Styles */
-.error {
-  color: #e74c3c;
-  font-size: 12px;
-  margin-top: 5px;
-  display: block;
-}
-
-.error-input {
-  border-color: #e74c3c !important;
-}
-
-/* Success Message Styles */
-.success-message {
-  background-color: #2ecc71;
-  color: white;
-  padding: 15px;
-  border-radius: 4px;
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-/* Submit button styles */
-.submit-btn {
-  background-color: #ff8e3a;
-  color: white;
-  padding: 12px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-weight: 600;
-  font-size: 16px;
-  margin-top: 10px;
-  width: 100%;
-  transition: background-color 0.3s;
-}
-
-.submit-btn:hover {
-  background-color: #e8771e;
-}
-
-.submit-btn:disabled {
-  background-color: #ccc;
-  cursor: not-allowed;
-}
-
-/* Responsive adjustments */
-@media (max-width: 600px) {
-  .popup-form {
-    padding: 20px;
-    max-width: 95%;
-  }
-  
-  .popup-form h3 {
-    font-size: 20px;
-  }
-  
-  .form-row {
-    flex-direction: column;
-    gap: 15px;
-  }
-  
-  .checkbox-row label {
-    font-size: 13px;
-  }
 }
